Update fade duration to float property

+++ include/openspace/engine/openspaceengine.h
@@ -30,7 +30,7 @@
 #include <openspace/properties/propertyowner.h>
 #include <openspace/properties/property.h>
 #include <openspace/properties/scalar/boolproperty.h>
-#include <openspace/properties/scalar/intproperty.h>
+#include <openspace/properties/scalar/floatproperty.h>
 #include <openspace/properties/stringproperty.h>
 #include <openspace/scene/profile.h>
 #include <openspace/util/keys.h>
@@ -146,7 +146,7 @@ private:
     properties::BoolProperty _printEvents;
     properties::OptionProperty _visibility;
     properties::BoolProperty _showHiddenSceneGraphNodes;
-    properties::IntProperty _fadeOnEnableDuration;
+    properties::FloatProperty _fadeOnEnableDuration;
     properties::BoolProperty _disableAllMouseInputs;
 
     std::unique_ptr<Scene> _scene;
